requirements.txt
# Core Dependencies
streamlit
python-dotenv

# Google ADK (CRITICAL - NEW!)
google-generativeai
google-adk
#google-adk-agents

# SEO Tools
requests
beautifulsoup4
lxml

# Data Models
pydantic
pydantic-settings

# Utilities
urllib3
certifi
PyYAML
